create separate room component, add form to update room
